Extract Choices component Extract the choices list to its own component.
