+++ app/classifier/tasks/survey/chooser/Chooser.jsx
@@ -2,28 +2,10 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import sortIntoColumns from 'sort-into-columns';
 import Translate from 'react-translate-component';
-import Thumbnail from '../../../../components/thumbnail';
 import CharacteristicsFilter from './CharacteristicsFilter';
-
-// key codes
-const BACKSPACE = 8;
-const UP = 38;
-const DOWN = 40;
+import Choices from './Choices';
 
 class Chooser extends React.Component {
-  constructor() {
-    super();
-    this.choiceButtons = [];
-    this.handleClearFilters = this.handleClearFilters.bind(this);
-  }
-
-  componentDidMount() {
-    this.sortChoiceButtons();
-  }
-
-  componentDidUpdate() {
-    this.sortChoiceButtons();
-  }
 
   getFilteredChoices() {
     return this.props.task.choicesOrder.map((choiceId) => {
@@ -53,70 +35,11 @@ class Chooser extends React.Component {
     }
   }
 
-  sortChoiceButtons() {
-    // overrides default DOM focus order by sorting the buttons according to task.choicesOrder
-    const newChoiceButtons = [];
-    this.choiceButtons
-      .filter(Boolean)
-      .map((button) => {
-        const choiceID = button.getAttribute('data-choiceID');
-        const index = this.props.task.choicesOrder.indexOf(choiceID);
-        newChoiceButtons[index] = button;
-      });
-    this.choiceButtons = newChoiceButtons.filter(Boolean);
-  }
-
-  whatSizeThumbnails({ length }) {
-    if (length <= 5) {
-      return 'large';
-    } else if (length <= 10) {
-      return 'medium';
-    } else if (length <= 20) {
-      return 'small';
-    } else if (this.props.task.alwaysShowThumbnails) {
-      return 'small';
-    } else {
-      return 'none';
-    }
-  }
-
-  handleClearFilters() {
-    this.props.task.characteristicsOrder.map(characteristicId => this.props.onFilter(characteristicId, undefined));
-  }
-
-  handleKeyDown(choiceId, e) {
-    const index = this.choiceButtons.indexOf(document.activeElement);
-    let newIndex;
-    switch (e.which) {
-      case BACKSPACE:
-        this.props.onRemove(choiceId);
-        e.preventDefault();
-        break;
-      case UP:
-        newIndex = index - 1;
-        if (newIndex === -1) {
-          newIndex = this.choiceButtons.length - 1;
-        }
-        this.choiceButtons[newIndex].focus();
-        e.preventDefault();
-        break;
-      case DOWN:
-        newIndex = (index + 1) % this.choiceButtons.length;
-        this.choiceButtons[newIndex].focus();
-        e.preventDefault();
-        break;
-      default:
-    }
-  }
-
   render() {
-    const { task, translation, filters, onFilter } = this.props;
-    this.choiceButtons = [];
+    const { annotation, task, translation, filters, focusedChoice, onChoose, onFilter, onRemove } = this.props;
     const filteredChoices = this.getFilteredChoices();
-    const thumbnailSize = this.whatSizeThumbnails(filteredChoices);
     const columnsCount = this.howManyColumns(filteredChoices);
     const sortedFilteredChoices = sortIntoColumns(filteredChoices, columnsCount);
-    const selectedChoices = this.props.annotation.value.map(item => item.choice);
     return (
       <div className="survey-task-chooser">
         <CharacteristicsFilter
@@ -127,52 +50,18 @@ class Chooser extends React.Component {
           onFilter={onFilter}
         />
         <hr className="survey-task-chooser__divider" />
-        <div className="survey-task-chooser-choices" data-thumbnail-size={thumbnailSize} data-columns={columnsCount}>
-          {sortedFilteredChoices.length === 0 && <div><em>No matches.</em></div>}
-          {sortedFilteredChoices.map((choiceId, i) => {
-            const choice = task.choices[choiceId];
-            const chosenAlready = selectedChoices.indexOf(choiceId) > -1;
-            let tabIndex = -1;
-            if (i === 0 && this.props.focusedChoice.length === 0) {
-              tabIndex = 0;
-            }
-            if (choiceId === this.props.focusedChoice) {
-              tabIndex = 0;
-            }
-            const src = this.props.task.images[choice.images[0]];
-            const srcPath = Thumbnail.getThumbnailSrc({
-              origin: 'https://thumbnails.zooniverse.org',
-              width: 500,
-              height: 500,
-              src
-            });
-            const thumbnail = srcPath || '';
-            return (
-              <button
-                autoFocus={choiceId === this.props.focusedChoice}
-                key={choiceId}
-                data-choiceID={choiceId}
-                ref={button => this.choiceButtons.push(button)}
-                tabIndex={tabIndex}
-                type="button"
-                className={`survey-task-chooser-choice-button ${chosenAlready ? 'survey-task-chooser-choice-button-chosen' : ''}`}
-                onClick={this.props.onChoose.bind(null, choiceId)}
-                onKeyDown={this.handleKeyDown.bind(this, choiceId)}
-              >
-                <span className="survey-task-chooser-choice">
-                  {choice.images.length > 0 &&
-                    <span
-                      className="survey-task-chooser-choice-thumbnail"
-                      role="presentation"
-                      style={{ backgroundImage: `url('${thumbnail}')` }}
-                    />
-                  }
-                  <span className="survey-task-chooser-choice-label">{translation.choices[choiceId].label}</span>
-                </span>
-              </button>
-            );
-          })}
-        </div>
+        <Choices
+          annotation={annotation}
+          columnsCount={columnsCount}
+          filters={filters}
+          filteredChoices={filteredChoices}
+          focusedChoice={focusedChoice}
+          sortedFilteredChoices={sortedFilteredChoices}
+          task={task}
+          translation={translation}
+          onChoose={onChoose}
+          onRemove={onRemove}
+        />
         <div style={{ textAlign: 'center' }}>
           <Translate
             content="tasks.survey.showing"
